Create file for survey handler in context
